Moved yara scan function to base_system
